backend/requirements.txt
# Django and API requirements
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.1.0
drf-spectacular==0.27.0
redis[hiredis]>=5.0.1

# Celery and Redis requirements
celery==5.3.5

# Development requirements
python-dotenv==1.0.0

setuptools>=65.5.0
daphne
